Check DocRef.type binding for US Core mandatory profile
It's a required binding that I missed earlier.
It's not required in base FHIR, so it needs to be checked by us here.

The official spec says that only a subset of LOINC is in the binding,
but without adding a full LOINC db to this project, that's difficult to
check. Instead, we just look for any LOINC code (or the UNK null
flavor).