q_ref_target_valid: change denominator from fields to rows
Before, we would count all potential matching Reference fields as the
denominator. This commit aligns us back with the qualifier metric
documentation to just use all rows as the denominator.

In cases like DocumentReference.context.encounter that can have
multiple references, we include the row once in the numerator and
concatenate the target values for debugging.