Add some missing US Core checks now that we have attachment info
Previously, the US Core checks around DiagnosticReport.presentedForm
and DocumentReference.content.attachment were marked as "quirks" that
we ignored due to the ETL stripping out the bits necessary to make
those checks.

But now that the ETL leaves us enough info, we don't need to mark
them as quirks and can just implement them.

This made the DocumentReference mandatory cube too large, however,
so it's now split into two tables (like Observations are).
So this is a breaking change for consumers, sorry.