fix: allow system-only codings
This was surfaced after we started stripping codes and displays from
certain Epic-internal codings. We would fail Encounter.class (which
is a Coding) on q_valid_us_core_v4 because we wouldn't consider it
filled in.

But... we were being more strict for Codings than we were for
CodeableConcepts, where we just required that any Codings were
defined underneath it at all, system-only or otherwise. Likewise,
the spec actually does suggest that system-only Codings are a real
and (kinda) meaningful state:

> If the system is present, and there is no code, then this is
> understood to mean that there is no suitable code in the system
> in which to represent the code.

So... let's allow it. This will fix our bogus failures at least.